musubisuite_back/requirements.txt
# Django Backend Requirements for MusubiSuite

# Core Framework
Django==5.0.0
djangorestframework==3.14.0

# Database
psycopg2-binary==2.9.9

# CORS Headers
django-cors-headers==4.3.1

# Authentication & Security
djangorestframework-simplejwt==5.3.1

# Filtering & Search
django-filter==23.5

# Environment Variables
python-decouple==3.8

# Date/Time utilities
python-dateutil==2.8.2

# API Documentation
drf-spectacular==0.27.0

# Production Server
gunicorn==21.2.0
whitenoise==6.6.0

# Utils
Pillow==10.1.0
